move mi tests from test_indexing_slow.py
